requirements.txt
pgmpy==0.1.25
pandas>=2.0
numpy>=1.24
rdflib>=7.0
scikit-learn>=1.4
pyyaml>=6.0
